updated current user context for multiple university assignments

+++ ScholarshipManagementAPI/DTOs/University/Courses/CourseRequestDto.cs
@@ -26,7 +26,7 @@ namespace ScholarshipManagementAPI.DTOs.University.Courses
 
         // Faculties Mapping
         public List<CourseFacultyDto> Faculties { get; set; } = new();
-        public List<int> FacultyIds { get; set; } = new();
+        public List<long> FacultyIds { get; set; } = new();
 
 
         // Response
